recrawler/requirements.txt
lxml
ujson
gevent
murmur
urlnorm
requests
cityhash
mysql-python
python-magic
